Reduce SPI buffer to a single buffer
The ISM43362 WIFI drivers works in a command / response mechanism
that does not allow buffering. For each command that is sent, an answer
is awaited until the next command will be accepted.

As a consequence, there is no need to allow a multple number of buffers
to be managed in BufferedSPI object.

Reducing tx_multiple to 1 here also will free up to 7500 bytes of RAM
in Heap.